refactor: replace convert word to swap
