+++ packages/core/src/controllers/AccountController.ts
@@ -5,8 +5,8 @@ import type { CaipAddress, ConnectedWalletInfo } from '../utils/TypeUtil.js'
 import type { Balance } from '@web3modal/common'
 import { BlockchainApiController } from './BlockchainApiController.js'
 import { SnackController } from './SnackController.js'
-import { ConvertController } from './ConvertController.js'
-import { ConvertApiUtil } from '../utils/ConvertApiUtil.js'
+import { SwapController } from './SwapController.js'
+import { SwapApiUtil } from '../utils/SwapApiUtil.js'
 import type { W3mFrameTypes } from '@web3modal/wallet'
 
 // -- Types --------------------------------------------- //
@@ -105,7 +105,7 @@ export const AccountController = {
         const response = await BlockchainApiController.getBalance(state.address)
 
         this.setTokenBalance(response.balances)
-        ConvertController.setBalances(ConvertApiUtil.mapBalancesToConvertTokens(response.balances))
+        SwapController.setBalances(SwapApiUtil.mapBalancesToSwapTokens(response.balances))
       }
     } catch (error) {
       SnackController.showError('Failed to fetch token balance')
