+++ packages/core/src/controllers/RouterController.ts
@@ -1,7 +1,7 @@
 import { subscribeKey as subKey } from 'valtio/vanilla/utils'
 import { proxy } from 'valtio/vanilla'
 import type { CaipNetwork, Connector, WcWallet } from '../utils/TypeUtil.js'
-import type { ConvertInputTarget } from './ConvertController.js'
+import type { SwapInputTarget } from './SwapController.js'
 
 // -- Types --------------------------------------------- //
 type TransactionAction = {
@@ -47,9 +47,9 @@ export interface RouterControllerState {
     | 'WhatIsANetwork'
     | 'WhatIsAWallet'
     | 'WhatIsABuy'
-    | 'Convert'
-    | 'ConvertSelectToken'
-    | 'ConvertPreview'
+    | 'Swap'
+    | 'SwapSelectToken'
+    | 'SwapPreview'
   history: RouterControllerState['view'][]
   data?: {
     connector?: Connector
@@ -57,7 +57,7 @@ export interface RouterControllerState {
     network?: CaipNetwork
     email?: string
     newEmail?: string
-    target?: ConvertInputTarget
+    target?: SwapInputTarget
   }
   transactionStack: TransactionAction[]
 }
