+++ packages/scaffold/index.ts
@@ -21,10 +21,10 @@ export * from './src/views/w3m-onramp-activity-view/index.js'
 export * from './src/views/w3m-onramp-fiat-select-view/index.js'
 export * from './src/views/w3m-onramp-providers-view/index.js'
 export * from './src/views/w3m-onramp-tokens-select-view/index.js'
-export * from './src/views/w3m-convert-view/index.js'
-export * from './src/views/w3m-convert-preview-view/index.js'
-export * from './src/views/w3m-convert-select-token-view/index.js'
-export * from './src/views/w3m-convert-view/index.js'
+export * from './src/views/w3m-swap-view/index.js'
+export * from './src/views/w3m-swap-preview-view/index.js'
+export * from './src/views/w3m-swap-select-token-view/index.js'
+export * from './src/views/w3m-swap-view/index.js'
 export * from './src/views/w3m-transactions-view/index.js'
 export * from './src/views/w3m-what-is-a-network-view/index.js'
 export * from './src/views/w3m-what-is-a-wallet-view/index.js'
@@ -53,11 +53,11 @@ export * from './src/partials/w3m-connecting-wc-mobile/index.js'
 export * from './src/partials/w3m-connecting-wc-qrcode/index.js'
 export * from './src/partials/w3m-connecting-wc-unsupported/index.js'
 export * from './src/partials/w3m-connecting-wc-web/index.js'
-export * from './src/partials/w3m-convert-details/index.js'
-export * from './src/partials/w3m-convert-input/index.js'
+export * from './src/partials/w3m-swap-details/index.js'
+export * from './src/partials/w3m-swap-input/index.js'
 export * from './src/partials/w3m-header/index.js'
 export * from './src/partials/w3m-help-widget/index.js'
-export * from './src/partials/w3m-swap-input/index.js'
+export * from './src/partials/w3m-onramp-input/index.js'
 export * from './src/partials/w3m-legal-footer/index.js'
 export * from './src/partials/w3m-mobile-download-links/index.js'
 export * from './src/partials/w3m-onramp-providers-footer/index.js'
