+++ packages/scaffold/src/modal/w3m-router/index.ts
@@ -117,12 +117,12 @@ export class W3mRouter extends LitElement {
         return html`<w3m-wallet-receive-view></w3m-wallet-receive-view>`
       case 'WalletCompatibleNetworks':
         return html`<w3m-wallet-compatible-networks-view></w3m-wallet-compatible-networks-view>`
-      case 'Convert':
-        return html`<w3m-convert-view></w3m-convert-view>`
-      case 'ConvertSelectToken':
-        return html`<w3m-convert-select-token-view></w3m-convert-select-token-view>`
-      case 'ConvertPreview':
-        return html`<w3m-convert-preview-view></w3m-convert-preview-view>`
+      case 'Swap':
+        return html`<w3m-swap-view></w3m-swap-view>`
+      case 'SwapSelectToken':
+        return html`<w3m-swap-select-token-view></w3m-swap-select-token-view>`
+      case 'SwapPreview':
+        return html`<w3m-swap-preview-view></w3m-swap-preview-view>`
       case 'WalletSend':
         return html`<w3m-wallet-send-view></w3m-wallet-send-view>`
       case 'WalletSendSelectToken':
