--- packages/scaffold/src/partials/w3m-convert-input/styles.ts
@@ -1,131 +1,0 @@
-import { css } from 'lit'
-
-export default css`
-  :host > wui-flex {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    border-radius: var(--wui-border-radius-s);
-    padding: var(--wui-spacing-xl);
-    padding-right: var(--wui-spacing-s);
-    width: 100%;
-    height: 100px;
-    box-sizing: border-box;
-    position: relative;
-  }
-
-  :host > wui-flex > svg.input_mask {
-    position: absolute;
-    inset: 0;
-    z-index: 5;
-  }
-
-  :host wui-flex .input_mask__border,
-  :host wui-flex .input_mask__background {
-    transition: fill var(--wui-duration-md) var(--wui-ease-out-power-1);
-    will-change: fill;
-  }
-
-  :host wui-flex .input_mask__border {
-    fill: var(--wui-gray-glass-005);
-  }
-
-  :host wui-flex .input_mask__background {
-    fill: var(--wui-gray-glass-002);
-  }
-
-  :host wui-flex.focus .input_mask__border {
-    fill: var(--wui-gray-glass-020);
-  }
-
-  :host > wui-flex .swap-input,
-  :host > wui-flex .swap-token-button {
-    z-index: 10;
-  }
-
-  :host > wui-flex .swap-input {
-    -webkit-mask-image: linear-gradient(
-      270deg,
-      transparent 0px,
-      transparent 8px,
-      black 24px,
-      black 25px,
-      black 32px,
-      black 100%
-    );
-    mask-image: linear-gradient(
-      270deg,
-      transparent 0px,
-      transparent 8px,
-      black 24px,
-      black 25px,
-      black 32px,
-      black 100%
-    );
-  }
-
-  :host > wui-flex .swap-input input {
-    background: none;
-    border: none;
-    height: 42px;
-    width: 100%;
-    font-size: 32px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 130%;
-    letter-spacing: -1.28px;
-    outline: none;
-    caret-color: var(--wui-color-accent-100);
-    color: var(--wui-color-fg-100);
-  }
-
-  :host > wui-flex .swap-input input:focus-visible {
-    outline: none;
-  }
-
-  :host > wui-flex .swap-input input::-webkit-outer-spin-button,
-  :host > wui-flex .swap-input input::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-
-  .token-select-button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: var(--wui-spacing-xxs);
-    padding: var(--wui-spacing-xs);
-    padding-right: var(--wui-spacing-1xs);
-    height: 40px;
-    border: none;
-    border-radius: 80px;
-    background: var(--wui-gray-glass-002);
-    box-shadow: inset 0 0 0 1px var(--wui-gray-glass-002);
-    cursor: pointer;
-    transition: background 0.2s linear;
-  }
-
-  .token-select-button:hover {
-    background: var(--wui-gray-glass-005);
-  }
-
-  .token-select-button wui-image {
-    width: 24px;
-    height: 24px;
-    border-radius: var(--wui-border-radius-s);
-    box-shadow: inset 0 0 0 1px var(--wui-gray-glass-010);
-  }
-
-  .max-value-button {
-    background-color: transparent;
-    border: none;
-    cursor: pointer;
-    color: var(--wui-gray-glass-020);
-    padding-left: 0px;
-  }
-
-  .market-value {
-    min-height: 18px;
-  }
-`
