+++ packages/scaffold/src/partials/w3m-onramp-input/styles.ts
@@ -1,0 +1,35 @@
+import { css } from 'lit'
+
+export default css`
+  :host {
+    width: 100%;
+  }
+
+  wui-loading-spinner {
+    position: absolute;
+    top: 50%;
+    right: 20px;
+    transform: translateY(-50%);
+  }
+
+  .currency-container {
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    right: var(--wui-spacing-1xs);
+    height: 40px;
+    padding: var(--wui-spacing-xs) var(--wui-spacing-1xs) var(--wui-spacing-xs)
+      var(--wui-spacing-xs);
+    min-width: 95px;
+    border-radius: var(--FULL, 1000px);
+    border: 1px solid var(--wui-gray-glass-002);
+    background: var(--wui-gray-glass-002);
+    cursor: pointer;
+  }
+
+  .currency-container > wui-image {
+    height: 24px;
+    width: 24px;
+    border-radius: 50%;
+  }
+`
